Altera o nome do módulo de manager.py para app.py e utiliza o command line interface do flask 2.2.
